backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
anthropic==0.49.0
pdfplumber==0.11.4
python-docx==1.1.2
python-dotenv==1.0.1
gunicorn==23.0.0
